pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "shopify-llm-store-creator"
version = "1.0.0"
description = "AI-powered Shopify store creator with web interface"
readme = "docs/README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ninjtheturtle", email = "nengjiali@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "flask>=2.0.0",
    "flask-cors>=3.0.0",
    "requests>=2.25.0",
    "python-dotenv>=0.19.0",
    "shopify-python-api>=12.0.0",
    "transformers>=4.20.0",
    "torch>=1.12.0",
    "datasets>=2.0.0",
    "peft>=0.3.0",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.39.0",
    "trl>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.900",
]

[project.scripts]
shopify-llm = "src.api.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
